chore: refactor plunker translators with a new systemjs one
